requirements.txt
streamlit
tensorflow==2.15.0
keras==3.10.0
pillow
numpy